optimizing code
