+++ src/test/java/helper/Util.java
@@ -5,6 +5,17 @@ import com.google.gson.JsonObject;
 import java.util.HashMap;
 
 public class Util {
+
+    private static String authToken;
+
+    public static void setAuthToken(String token) {
+        authToken = token;
+    }
+
+    public static String getAuthToken() {
+        return authToken;
+    }
+
     /**
      * returns a json body as string for login request
      * @param username
