--- src/test/java/runner/common/Base.java
@@ -1,13 +1,0 @@
-package runner.common;
-
-public class Base {
-    private static String authToken;
-
-    public static void setAuthToken(String token) {
-        authToken = token;
-    }
-
-    public static String getAuthToken() {
-        return authToken;
-    }
-}
